Solved (or partially solved, as there is more than one solver implementation and I don't have time to look at all of them) a bug in master version. Detail explaination:
The solver cache does not depend on execution state, i.e. all execution states share one solver. The solver cache is built as a map, with keys to the map a vector of constraints.

The keys itself are not ideal as well, which will be explained later. Without optimization this implementation is correct (however slow). The default optimization is to reduce the constraints to an independent set of constrains that only relates to the expression that is relevant. This creates problem, an example is as follows:

One set of symbolic expressions: a, b, c

constraints are: 0<a<5, 1<b<6.
For one execution path, we may have them solved as a=1,b=2,c=3.
On another execution path we may have them solved as a=4,b=5. (c does not matter here)

Assume the second execution path is first evaluated, thus for constraints 0<a<5, 1<b<6, the cache maps it to values a = 4, b = 5.

Then on the first execution path, assume it is executed after the second execution path, a value d is computed as a+b+c, and c has some constraint c == 3. The constraint now is then 0<a<5, 1<b<6, c == 3. The solver now solves it as a = 1, b = 2, c = 3.
Then on the same execution path, a query with only a and b is now encountered, for example e = a+c; klee_get_valued(e);

In this case, a optimized query that relates to e, is :0<a<5, 1<b<6.
This query matches to the second execution path, which causes a and b to have inconsistent values.

This bug does not show often because without randomness (or some other reason), the solver always give the same resolution of a and b no matter constraint 1 or constraint 2 is encountered. However in my case the solver returns different (but valid) solution to two different constraints. I may need to look at the solver code to see why.

OPTIMIZATION AVAILABLE:
one optimization available is that now, the cache is simply a map. When constraints are out of order of each other, even though they may represent the same set of constraints, it is still a cache miss. Further more, we should be able to utilize caches that have constraints as super set of the query constraints. However I am not sure if it is worth it to do set computations. If there is librarys that can do it efficiently, I would recommend add it to the MapOfSets.h.

Gladtbx