Add LeakSanitizer (component of AddressSanitizer) suppression file which is required to suppress all the leaks I'm currently seeing in KLEE when running ``make unittests`` and ``make check``.
Ideally there should be no leaks but we aren't there yet. Hopefully
at some point we won't need to suppress any leaks and then we can
have a TravisCI build that builds with ASan.

The leak of the expression objects when running the executor is worrying
and I will investigate this next.