Allow relocation of installed klee tree
If klee is configured with certain bindir and runtime dir,
allow klee to be relocated, as long as subdirectory structure remains
intact.

For example, if klee is configured with bindir /usr/bin, and with
runtime dir /usr/lib/klee, but is relocated to certain directory
$RDIR, then running $RDIR/usr/bin/klee will search for runtime libraries
in $RDIR/usr/lib/klee.

Klee will use global runtime directory only when installed to global
binary directory.

Inspired by relocation code in gcc.